backend/requirements.txt
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
distro==1.9.0
dydantic==0.0.8
executing==2.2.0
fastapi==0.115.14
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-anthropic==0.3.17
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langmem==0.0.27
langsmith==0.4.4
matplotlib-inline==0.1.7
openai==1.93.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
traitlets==5.14.3
trustcall==0.0.39
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.23.0
